[ASTReader] Merge ObjC interface/protocol definitions deterministically.
It's possible for a user to get into a situation where two definitions
of the same ObjC decl are loaded (due to a modular and non-modular
include of the same header, say). Previously (in 2ba19793512) some
effort was put into keeping the DefinitionData "invariant" in such cases,
at least when loading duplicate ObjC interfaces.

Unfortunately, this still had two shortcomings:

  1. It only applied to ObjCInterfaceDecls, not ObjCProtocolDecls.
  2. It was order-sensitive: whichever defn was loaded first "stuck"
     and subsequent defns would be erased / overwritten by it.

We therefore extend the logic to cover protocols and add to it a
step that "determinizes" the winning definition, favouring modular
definitions over non (favouring higher-numbered module IDs in general).